requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
cloudinary==1.44.1
fastapi==0.128.0
h11==0.14.0
httpcore==0.16.3
httptools==0.7.1
httpx==0.23.3
idna==3.11
innertube==2.1.19
maturin==1.11.5
mediate==0.1.8
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
rfc3986==1.5.0
roster==0.1.11
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.45.1
yt-dlp==2025.12.8
